% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCytoTRACE.R
\name{iCytoTRACE}
\alias{iCytoTRACE}
\title{integrated CytoTRACE (iCytoTRACE)}
\usage{
iCytoTRACE(datasets, enableFast = TRUE, ncores = 1, subsamplesize = 1000)
}
\arguments{
\item{datasets}{list of gene expression matrices where columns are cells and rows are genes}
}
\value{
a list containing
\itemize{
\item exprMatrix: a matrix (genes by single cells) of the corrected gene expression values from Scanorama
\item CytoTRACE: a numeric vector of the predicted ordering of single cells from 1.0 (least differentiated) to 0.0 (most differentiated)
\item CytoTRACErank: a numeric vector of the ranked predicted ordering of single cells by differentiation status. High ranks correspond to less differentiated cells, while low ranks correspond to more differentiated cells.
\item GCS: a numeric vector of the merged gene counts signature (geometric mean of the top 200 genes associated with gene counts)
\item Counts: a numeric vector of the corrected number of genes expressed per single cell (gene counts)
\item coord: a matrix containing the coordinates for the merged low-dimensional embedding (number of cells x 100 t-SNE components)
\item filteredCells = a character vector of the names of single cells (columns) that were filtered due to poor quality.
}
}
\description{
This function generates single-cell predictions of differentiation status
across multiple, heterogeneous scRNA-seq batches/datasets. This implementation leverages
mutual nearest neighbor and Gaussian kernel normalization techniques from Scanorama to merge datasets.
It takes in a list of gene expression matrices where columns are cells and rows are genes and outputs
integrated gene counts, gene counts signature (GCS), CytoTRACE, and the correct gene expression matrix.
}
\examples{

#Create a list containing two bone marrow scRNA-seq datasets profiled on different platforms, 10x and Smart-seq2
datasets <- list(marrow_10x_expr, marrow_plate_expr)

#Run iCytoTRACE
results <- iCytoTRACE(datasets)
}
\references{
https://doi.org/10.1101/649848
}
\seealso{
https://cytotrace.stanford.edu
}
\author{
Gunsagar Gulati <cytotrace@gmail.com>
}
