% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{marrow_plate_expr}
\alias{marrow_plate_expr}
\title{Mouse bone marrow differentiation scRNA-seq (10x) expression dataset}
\format{
A data frame where single cells are columns and genes are rows.
Values represent raw counts.
}
\source{
GSE109774 and \href{https://github.com/czbiohub/tabula-muris}{Tabula Muris GitHub}
}
\usage{
marrow_plate_expr
}
\description{
Single-cell RNA-sequencing data from Tabula Muris of mouse bone marrow cells
profiled using the Smart-seq2 platform.
}
\references{
Schaum, Nicholas, et al. "Single-cell transcriptomics of 20 mouse organs creates a Tabula Muris: The Tabula Muris Consortium." Nature 562.7727 (2018): 367.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/30283141}{PubMed})
}
\keyword{RNA-sequencing,}
\keyword{Smart-seq2,}
\keyword{bone}
\keyword{expression}
\keyword{gene}
\keyword{marrow,}
\keyword{single-cell}
