% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{marrow_plate_pheno}
\alias{marrow_plate_pheno}
\title{Mouse bone marrow differentiation scRNA-seq (Smart-seq2) phenotype labels}
\format{
A character vector of the phenotype labels of mouse bone marrow cells with names corresponding to the single cell IDs
}
\source{
GSE109774 and \href{https://github.com/czbiohub/tabula-muris}{Tabula Muris GitHub}
}
\usage{
marrow_plate_pheno
}
\description{
Phenotype labels of mouse bone marrow cells from Tabula Muris corresponding to the data file 'marrow_10x_expr'.
}
\references{
Schaum, Nicholas, et al. "Single-cell transcriptomics of 20 mouse organs creates a Tabula Muris: The Tabula Muris Consortium." Nature 562.7727 (2018): 367.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/30283141}{PubMed})
}
\keyword{RNA-sequencing,}
\keyword{Smart-seq2,}
\keyword{bone}
\keyword{marrow,}
\keyword{phenotype}
\keyword{single-cell}
