% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoTRACE.R
\name{plotCytoTRACE}
\alias{plotCytoTRACE}
\title{Visualizing CytoTRACE results}
\usage{
plotCytoTRACE(
  cyto_obj = NULL,
  phenotype = NULL,
  gene = NULL,
  colors = NULL,
  emb = NULL,
  otherName = NULL,
  otherValue = NULL,
  outputDir = "./"
)
}
\arguments{
\item{cyto_obj}{a list object generated by the 'CytoTRACE' or 'iCytoTRACE' functions}

\item{phenotype}{(optional) a character vector of phenotype labels corresponding to the single cells in 'cyto_obj'. The vector names should match the column names of the original gene expression matrix.}

\item{gene}{(optional) a case-insensitive string corresponding to a gene of interest. The case-insensitive string should match values provided as row names of the original gene expression matrix.}

\item{colors}{(optional) a character vector of colors to be used to create a palette for the plots}

\item{emb}{(optional) a data frame containing two columns corresponding to the 2D embeddings and where the row names match the column names of the original gene expression matrix.}

\item{otherName}{(optional) a string to replace 'gene' or describe the input values in 'otherValue'}

\item{otherValue}{(optional) a numeric vector of values to plot with CytoTRACE. This will replace the 'gene' plot.}

\item{outputDir}{path to output directory where plots and tables will be written. Format should be "/PATH/TO/DIRECTORY/".}
}
\value{
a PDF of 2D embedded plots colored by CytoTRACE, and, if provided, phenotype labels, and gene expression.

a PDF of a boxplot of CytoTRACE by phenotype labels (if provided).

a tab-delimited text file containing a table of CytoTRACE values t-SNE embeddings, and, if provided, phenotype labels and gene expression values.
}
\description{
This function generates 2D plots and tables to visualize CytoTRACE, known phenotypes, and gene expression.
The current implementation uses t-SNE for dimensional reduction but users can also input their own embeddings.
At minimum, the ‘plotCytoTRACE’ function takes as input a list object generated by either the ‘CytoTRACE’ or ‘iCytoTRACE’ functions.
Users can also optionally provide phenotype labels or gene names to generate additional plots.
Boxplots of CytoTRACE by phenotype labels are automatically generated when phenotype labels are provided.

cyto_obj = NULL, phenotype = NULL, gene = NULL, colors = NULL, emb = NULL
}
\examples{

#Use the bone marrow 10x scRNA-seq dataset to run CytoTRACE
results <- CytoTRACE(marrow_10x_expr)

#Run plotCytoTRACE
plotCytoTRACE(results, phenotype = marrow_10x_pheno, gene = "Kit")

}
\references{
https://doi.org/10.1101/649848
}
\seealso{
https://cytotrace.stanford.edu
}
\author{
Gunsagar Gulati <cytotrace@gmail.com>
}
