% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoTRACE.R
\name{CytoTRACE}
\alias{CytoTRACE}
\title{CytoTRACE}
\usage{
CytoTRACE(
  mat,
  batch = NULL,
  enableFast = TRUE,
  ncores = 1,
  subsamplesize = 1000
)
}
\arguments{
\item{mat}{matrix of gene expression values where columns are cells and rows are genes}

\item{batch}{character vector of length equal to the number of columns (cells) in the matrix}

\item{enableFast}{boolean indicating whether or not to run CytoTRACE in 'fast mode' for datasets with >3,000 cells. Fast mode uses a subsampling approach to reduce runtime.}

\item{ncores}{integer indicating the number of cores to utilize when enableFast = TRUE}

\item{subsamplesize}{integer indicating the number of cells to subsample when enableFast = TRUE}
}
\value{
a list containing
\itemize{
\item CytoTRACE: a numeric vector of the predicted ordering of single cells from 1.0 (least differentiated) to 0.0 (most differentiated)
\item CytoTRACErank: a numeric vector of the ranked predicted ordering of single cells by differentiation status. High ranks correspond to less differentiated cells, while low ranks correspond to more differentiated cells.
\item cytoGenes: a numeric vector of the Pearson correlation of each gene with CytoTRACE, ordered from highest to lowest
\item GCS: a numeric vector of the gene counts signature (geometric mean of the top 200 genes associated with gene counts)
\item gcsGenes: a numeric vector of the Pearson correlation of each gene with gene counts, ordered from highest to lowest
\item Counts: a numeric vector of the number of genes expressed per single cell (gene counts)
\item filteredCells: a character vector of the names of single cells (columns) that were filtered due to poor quality.
\item exprMatrix: a matrix of gene expression values after the following normalization steps: (i) sequencing depth normalization by rescaling single-cell transcriptomes to transcripts per million (TPM) or counts per million (CPM) (ii) log2-normalization with a pseudo-count of 1, and (iii) Census normalization (Qiu et al., 2017) to convert the gene expression matrix to relative transcript counts by rescaling single-cell transcriptomes to the total number of detectably expressed genes in that cell.

}
}
\description{
This function generates single-cell predictions of differentiation status from scRNA-seq data.
It takes in a matrix of gene expression values where columns are cells and rows are genes.\cr\cr
(Optional) Technical batches can be corrected with ComBat by inputting a character vector of batch
annotations matching the number of columns (cells) in the matrix. Furthermore, for the analysis of
large datasets (>3,000 cells), users can increase the speed performance by enabling a subsampling approach for calculation
and using multiple cores.
}
\examples{
#Use the bone marrow 10x scRNA-seq dataset to run CytoTRACE
results <- CytoTRACE(marrow_10x_expr)

#Run this dataset on fast mode using 1 core
results <- CytoTRACE(marrow_10x_expr, enableFast = TRUE, ncores = 1)

}
\references{
https://doi.org/10.1101/649848
}
\seealso{
https://cytotrace.stanford.edu
}
\author{
Gunsagar Gulati <cytotrace@gmail.com>
}
